use jwt-extended

+++ resources/item.py
@@ -1,6 +1,6 @@
 from models.item import ItemModel
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import get_jwt_identity, jwt_required, get_jwt
 import sqlite3
 class Item(Resource):
     parser = reqparse.RequestParser()
@@ -15,7 +15,7 @@ class Item(Resource):
         return {'message': 'Item not found'},404
     
    
-        
+    @jwt_required(fresh=True)
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {"message": f"An item with name '{name}' already exists"},400
@@ -31,8 +31,11 @@ class Item(Resource):
         
         return item.json(),201
     
-
+    @jwt_required()
     def delete(self,name):
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {"message":"Admin privileges required"}, 401
         item= ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -54,6 +57,10 @@ class Item(Resource):
   
     
 class ItemList(Resource):
+    @jwt_required(optional=True)
     def get(self):
-       
-        return {'items': [x.json() for x in ItemModel.query.all()]}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.query.all()]
+        if user_id:
+            return {'items': items},200
+        return {'items': [item['name'] for item in items], 'message':'More data available if you login'},200
